refactor: remove duplicate function
