+++ app/(inferenceai)/inferenceai/components/nav.tsx
@@ -2,7 +2,7 @@
 
 import Image from "next/image";
 import Link from "next/link";
-import { cn } from "@/app/utils/tailwind";
+import { cn } from "@/lib/utils";
 import { useState } from "react";
 import {
   NavigationMenu,
