+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { GoogleAnalytics } from "@next/third-parties/google";
 import "./globals.css";
 import data from "@/data.json";
-import { cn } from "@/app/utils/tailwind";
+import { cn } from "@/lib/utils";
 import Console from "@/app/components/console";
 import Script from "next/script";
 import { Toaster } from "@/components/ui/sonner";
